fastapi/requirements.txt
# Core FastAPI dependencies
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
httpx==0.27.2

# Rate limiting
slowapi==0.1.9

# Environment management
python-dotenv==1.1.1

# Performance optimization
gunicorn==21.2.0

# Development dependencies (optional - can be removed for production)
pytest==8.4.2
pytest-asyncio==1.2.0
hypothesis==6.122.3

pytest-cov==6.0.0
